refactor: transaction modal files structure
