+++ features/withdrawals/claim/transaction-modal-claim/index.ts
@@ -1,0 +1,1 @@
+export * from './transaction-modal-claim';
