--- features/withdrawals/claim/tx-modal/index.ts
@@ -1,1 +1,0 @@
-export * from './tx-claim-modal';
