--- features/withdrawals/request/tx-modal/index.ts
@@ -1,1 +1,0 @@
-export * from './tx-request-modal';
