--- features/withdrawals/shared/tx-stage-modal/index.ts
@@ -1,2 +1,0 @@
-export * from './tx-stage-modal';
-export * from './stages';
