--- shared/components/tx-stage-modal-content/index.ts
@@ -1,1 +1,0 @@
-export * from './tx-stage-modal-content';
