+++ shared/transaction-modal/stages/icons.tsx
