+++ shared/transaction-modal/stages/iconsStyles.ts
