+++ shared/transaction-modal/stages/styles.ts
