+++ shared/transaction-modal/stages/tx-stage-pending.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 
 import { TxLinkEtherscan } from 'shared/components/tx-link-etherscan';
-import { TxStageModalContent } from 'shared/components/tx-stage-modal-content';
+import { TransactionModalContent } from 'shared/transaction-modal/transaction-modal-content';
 import { StageIconBlock } from './icons';
 
 type TxStagePendingProps = {
@@ -14,7 +14,7 @@ export const TxStagePending: FC<TxStagePendingProps> = (props) => {
   const { title, description, txHash } = props;
 
   return (
-    <TxStageModalContent
+    <TransactionModalContent
       icon={<StageIconBlock />}
       title={title}
       description={description || 'Awaiting block confirmation'}
