+++ shared/transaction-modal/stages/tx-stage-permit.tsx
@@ -1,11 +1,11 @@
 import { FC } from 'react';
 
-import { TxStageModalContent } from 'shared/components/tx-stage-modal-content';
+import { TransactionModalContent } from 'shared/transaction-modal/transaction-modal-content';
 import { StageIconSign } from './icons';
 
 export const TxStagePermit: FC = () => {
   return (
-    <TxStageModalContent
+    <TransactionModalContent
       icon={<StageIconSign />}
       title="Please sign the message"
       description="Processing your request"
