+++ shared/transaction-modal/stages/tx-stage-success-multisig.tsx
@@ -1,9 +1,9 @@
-import { TxStageModalContent } from 'shared/components/tx-stage-modal-content';
+import { TransactionModalContent } from 'shared/transaction-modal/transaction-modal-content';
 import { StageIconSuccess } from './icons';
 
 export const TxStageSuccessMultisig = () => {
   return (
-    <TxStageModalContent
+    <TransactionModalContent
       icon={<StageIconSuccess />}
       title="Success"
       description="Your transaction has been successfully created in the multisig wallet and awaits approval from other participants"
