+++ shared/transaction-modal/transaction-modal-stages-routed/index.ts
@@ -1,0 +1,1 @@
+export * from './transaction-modal-stages-routed';
