+++ shared/transaction-modal/transaction-modal-stages-routed/styles.tsx
